# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habibi API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlatformStats {
  totalEndpoints: number;
  activeEndpoints: number;
  totalRequests: number;
  requestsToday: number;
  totalApikeys: number;
  uptime: string;
  creator: string;
  version: string;
}

export type EndpointItemStatus =
  (typeof EndpointItemStatus)[keyof typeof EndpointItemStatus];

export const EndpointItemStatus = {
  active: "active",
  maintenance: "maintenance",
  error: "error",
} as const;

export interface EndpointItem {
  name: string;
  path: string;
  method: string;
  category: string;
  description: string;
  params: string[];
  status: EndpointItemStatus;
  requiresApikey: boolean;
}

export interface EndpointList {
  total: number;
  categories: string[];
  endpoints: EndpointItem[];
}

export interface GenerateApikeyBody {
  name: string;
  email: string;
}

export interface ApikeyInfo {
  apikey: string;
  name: string;
  email: string;
  createdAt: string;
  requestCount: number;
  isValid: boolean;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type CheckApikeyParams = {
  apikey: string;
};