# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habibi API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get platform statistics
 */
export const GetStatsResponse = zod.object({
  totalEndpoints: zod.number(),
  activeEndpoints: zod.number(),
  totalRequests: zod.number(),
  requestsToday: zod.number(),
  totalApikeys: zod.number(),
  uptime: zod.string(),
  creator: zod.string(),
  version: zod.string(),
});

/**
 * @summary List all available endpoints
 */
export const ListEndpointsResponse = zod.object({
  total: zod.number(),
  categories: zod.array(zod.string()),
  endpoints: zod.array(
    zod.object({
      name: zod.string(),
      path: zod.string(),
      method: zod.string(),
      category: zod.string(),
      description: zod.string(),
      params: zod.array(zod.string()),
      status: zod.enum(["active", "maintenance", "error"]),
      requiresApikey: zod.boolean(),
    }),
  ),
});

/**
 * @summary Generate a new API key
 */
export const GenerateApikeyBody = zod.object({
  name: zod.string(),
  email: zod.string(),
});

export const GenerateApikeyResponse = zod.object({
  apikey: zod.string(),
  name: zod.string(),
  email: zod.string(),
  createdAt: zod.string(),
  requestCount: zod.number(),
  isValid: zod.boolean(),
});

/**
 * @summary Check API key validity and usage
 */
export const CheckApikeyQueryParams = zod.object({
  apikey: zod.coerce.string(),
});

export const CheckApikeyResponse = zod.object({
  apikey: zod.string(),
  name: zod.string(),
  email: zod.string(),
  createdAt: zod.string(),
  requestCount: zod.number(),
  isValid: zod.boolean(),
});