# lib/api-zod/src/generated/types/endpointItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habibi API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EndpointItemStatus } from "./endpointItemStatus";

export interface EndpointItem {
  name: string;
  path: string;
  method: string;
  category: string;
  description: string;
  params: string[];
  status: EndpointItemStatus;
  requiresApikey: boolean;
}