# lib/api-zod/src/generated/types/endpointList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habibi API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EndpointItem } from "./endpointItem";

export interface EndpointList {
  total: number;
  categories: string[];
  endpoints: EndpointItem[];
}