# lib/api-zod/src/generated/types/generateApikeyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habibi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateApikeyBody {
  name: string;
  email: string;
}