# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Habibi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalEndpoints: number;
  activeEndpoints: number;
  totalRequests: number;
  requestsToday: number;
  totalApikeys: number;
  uptime: string;
  creator: string;
  version: string;
}